+++ components/inputs/test/input-date-range.test.js
@@ -1,6 +1,6 @@
-import '../input-date-range.js';
 import { aTimeout, expect, fixture, oneEvent } from '@open-wc/testing';
 import { getDocumentLocaleSettings } from '@brightspace-ui/intl/lib/common.js';
+import { getShiftedEndDate } from '../input-date-range.js';
 import { runConstructor } from '../../../tools/constructor-test-helper.js';
 
 const basicFixture = '<d2l-input-date-range label="label text"></d2l-input-date-range>';
@@ -28,6 +28,55 @@ describe('d2l-input-date-range', () => {
 			runConstructor('d2l-input-date-range');
 		});
 
+	});
+
+	describe('utility', () => {
+		describe('getShiftedEndDate', () => {
+			it('should return correctly forward shifted end date if valid inputs', () => {
+				const start = '2020-10-26T04:00:00.000Z';
+				const end = '2020-10-27T04:00:00.000Z';
+				const prevStartValue = '2020-10-25T04:00:00.000Z';
+				const newEndValue = '2020-10-28T04:00:00.000Z';
+				expect(getShiftedEndDate(start, end, prevStartValue)).to.equal(newEndValue);
+			});
+
+			it('should return correctly backward shifted end date if valid inputs', () => {
+				const start = '2020-10-24T04:00:00.000Z';
+				const end = '2020-10-27T04:00:00.000Z';
+				const prevStartValue = '2020-10-25T04:00:00.000Z';
+				const newEndValue = '2020-10-26T04:00:00.000Z';
+				expect(getShiftedEndDate(start, end, prevStartValue)).to.equal(newEndValue);
+			});
+
+			it('should return correctly shifted end date if shift causes end date to go to next day', () => {
+				const start = '2020-10-25T11:00:00.000Z';
+				const end = '2020-10-27T23:30:00.000Z';
+				const prevStartValue = '2020-10-25T04:00:00.000Z';
+				const newEndValue = '2020-10-28T06:30:00.000Z';
+				expect(getShiftedEndDate(start, end, prevStartValue)).to.equal(newEndValue);
+			});
+
+			it('should return initial end date if prev start value was after end date', () => {
+				const start = '2020-10-23T04:00:00.000Z';
+				const end = '2020-10-22T04:00:00.000Z';
+				const prevStartValue = '2020-10-25T04:00:00.000Z';
+				expect(getShiftedEndDate(start, end, prevStartValue)).to.equal(end);
+			});
+
+			it('should return initial end date if not inclusive and prev start value was equal to end date', () => {
+				const start = '2020-10-20T04:00:00.000Z';
+				const end = '2020-10-25T04:00:00.000Z';
+				const prevStartValue = '2020-10-25T04:00:00.000Z';
+				expect(getShiftedEndDate(start, end, prevStartValue)).to.equal(end);
+			});
+
+			it('should return correctly shifted end date if inclusive and prev start value was equal to end date', () => {
+				const start = '2020-10-20T04:00:00.000Z';
+				const end = '2020-10-25T04:00:00.000Z';
+				const prevStartValue = '2020-10-25T04:00:00.000Z';
+				expect(getShiftedEndDate(start, end, prevStartValue, true)).to.equal(start);
+			});
+		});
 	});
 
 	describe('values', () => {
