+++ components/inputs/test/input-date-time-range.test.js
@@ -1,6 +1,6 @@
-import '../input-date-time-range.js';
 import { aTimeout, expect, fixture, oneEvent } from '@open-wc/testing';
 import { getDocumentLocaleSettings } from '@brightspace-ui/intl/lib/common.js';
+import { getShiftedEndDateTime } from '../input-date-time-range.js';
 import { runConstructor } from '../../../tools/constructor-test-helper.js';
 
 const basicFixture = '<d2l-input-date-time-range label="label text"></d2l-input-date-time-range>';
@@ -31,6 +31,178 @@ describe('d2l-input-date-time-range', () => {
 			runConstructor('d2l-input-date-time-range');
 		});
 
+	});
+
+	describe('utility', () => {
+		['America/Toronto', 'Australia/Eucla'].forEach((timezone) => {
+			const inclusive = false;
+
+			describe(`getShiftedEndDateTime in ${timezone}`, () => {
+
+				beforeEach(() => {
+					documentLocaleSettings.timezone.identifier = timezone;
+				});
+
+				afterEach(() => {
+					documentLocaleSettings.timezone.identifier = 'America/Toronto';
+				});
+
+				it('should return correctly forward shifted end date when localized without Z', () => {
+					const prevStartValue = '2020-10-25T04:00:00.000';
+					const prevEnd = '2020-10-27T04:00:00.000';
+					const start = '2020-10-26T04:00:00.000';
+					const newEndValue = '2020-10-28T04:00:00.000';
+					expect(getShiftedEndDateTime(start, prevEnd, prevStartValue, inclusive, true)).to.equal(newEndValue);
+				});
+
+				it('should shift as expected when times are different but dates the same because of UTC', () => {
+					let prevStartValue, start, prevEnd, newEndValue;
+
+					if (timezone === 'America/Toronto') {
+						prevStartValue = '2020-10-24T12:00:00.000Z';
+						prevEnd = '2020-10-25T02:00:00.000Z';
+						start = '2020-10-24T08:00:00.000Z';
+						newEndValue = '2020-10-24T22:00:00.000Z';
+					} else {
+						prevStartValue = '2020-10-23T22:00:00.000Z';
+						prevEnd = '2020-10-24T12:00:00.000Z';
+						start = '2020-10-23T18:00:00.000Z';
+						newEndValue = '2020-10-24T08:00:00.000Z';
+					}
+					expect(getShiftedEndDateTime(start, prevEnd, prevStartValue, inclusive, false)).to.equal(newEndValue);
+				});
+
+				it('should not shift if start and end dates are different because of UTC time', () => {
+					const prevStartValue = '2020-10-20T02:00:00.000Z';
+					const start = '2020-10-20T03:00:00.000Z';
+					let prevEnd;
+					if (timezone === 'America/Toronto') {
+						prevEnd = '2020-10-20T06:00:00.000Z';
+					} else {
+						prevEnd = '2020-10-20T20:00:00.000Z';
+					}
+					expect(getShiftedEndDateTime(start, prevEnd, prevStartValue, inclusive, false)).to.equal(prevEnd);
+				});
+
+			});
+
+			[true, false].forEach((localized) => {
+				describe(`getShiftedEndDateTime in ${timezone} with localized ${localized}`, () => {
+
+					beforeEach(() => {
+						documentLocaleSettings.timezone.identifier = timezone;
+					});
+
+					afterEach(() => {
+						documentLocaleSettings.timezone.identifier = 'America/Toronto';
+					});
+
+					describe('date change', () => {
+						it('should return correctly forward shifted end date', () => {
+							const prevStartValue = '2020-10-25T04:00:00.000Z';
+							const prevEnd = '2020-10-27T04:00:00.000Z';
+							const start = '2020-10-26T04:00:00.000Z';
+							const newEndValue = `2020-10-28T04:00:00.000${localized ? '' : 'Z'}`;
+							expect(getShiftedEndDateTime(start, prevEnd, prevStartValue, inclusive, localized)).to.equal(newEndValue);
+						});
+
+						it('should return correctly backward shifted end date', () => {
+							const prevStartValue = '2020-10-25T04:00:00.000Z';
+							const prevEnd = '2020-10-27T04:00:00.000Z';
+							const start = '2020-10-24T04:00:00.000Z';
+							const newEndValue = `2020-10-26T04:00:00.000${localized ? '' : 'Z'}`;
+							expect(getShiftedEndDateTime(start, prevEnd, prevStartValue, inclusive, localized)).to.equal(newEndValue);
+						});
+
+						it('should return correctly shifted end date when initial dates were equal', () => {
+							const prevStartValue = '2020-12-27T04:01:00.000Z';
+							const prevEnd = '2020-12-27T12:00:00.000Z';
+							const start = '2021-01-03T04:01:00.000Z';
+							const newEndValue = `2021-01-03T12:00:00.000${localized ? '' : 'Z'}`;
+							expect(getShiftedEndDateTime(start, prevEnd, prevStartValue, inclusive, localized)).to.equal(newEndValue);
+						});
+
+						it('should return initial end date if prev start value was after end date', () => {
+							const prevStartValue = '2020-10-25T04:00:00.000Z';
+							const prevEnd = '2020-10-22T04:00:00.000Z';
+							const start = '2020-10-21T04:00:00.000Z';
+							expect(getShiftedEndDateTime(start, prevEnd, prevStartValue, inclusive, localized)).to.equal(prevEnd);
+						});
+
+						it('should return initial end date if not inclusive and prev start value was equal to end date', () => {
+							const prevStartValue = '2020-10-25T04:00:00.000Z';
+							const prevEnd = '2020-10-25T04:00:00.000Z';
+							const start = '2020-10-20T04:00:00.000Z';
+							expect(getShiftedEndDateTime(start, prevEnd, prevStartValue, inclusive, localized)).to.equal(prevEnd);
+						});
+
+						it('should return correctly shifted end date if inclusive and prev start value was equal to end date', () => {
+							const prevStartValue = '2020-10-25T04:00:00.000Z';
+							const prevEnd = '2020-10-25T04:00:00.000Z';
+							const start = `2020-10-20T04:00:00.000${localized ? '' : 'Z'}`;
+							expect(getShiftedEndDateTime(start, prevEnd, prevStartValue, true, localized)).to.equal(start);
+						});
+					});
+
+					describe('time change', () => {
+						it('should not shift if start and end dates are different', () => {
+							const prevStartValue = '2020-10-20T06:00:00.000Z';
+							const prevEnd = '2020-10-25T06:00:00.000Z';
+							const start = '2020-10-20T08:00:00.000Z';
+							expect(getShiftedEndDateTime(start, prevEnd, prevStartValue, inclusive, localized)).to.equal(prevEnd);
+						});
+
+						it('should shift forward as expected when times are different but dates the same', () => {
+							const prevStartValue = '2020-10-25T06:00:00.000Z';
+							const prevEnd = '2020-10-25T07:00:00.000Z';
+							const start = '2020-10-25T08:00:00.000Z';
+							const newEndValue = `2020-10-25T09:00:00.000${localized ? '' : 'Z'}`;
+							expect(getShiftedEndDateTime(start, prevEnd, prevStartValue, inclusive, localized)).to.equal(newEndValue);
+						});
+
+						it('should shift backward as expected when times are different but dates the same', () => {
+							const prevStartValue = '2020-10-25T06:00:00.000Z';
+							const prevEnd = '2020-10-25T07:00:00.000Z';
+							const start = '2020-10-25T04:00:00.000Z';
+							const newEndValue = `2020-10-25T05:00:00.000${localized ? '' : 'Z'}`;
+							expect(getShiftedEndDateTime(start, prevEnd, prevStartValue, inclusive, localized)).to.equal(newEndValue);
+						});
+
+						it('should not shift when previously same and not inclusive', () => {
+							const prevStartValue = '2020-10-25T04:00:00.000Z';
+							const prevEnd = '2020-10-25T04:00:00.000Z';
+							const start = '2020-10-25T08:00:00.000Z';
+							expect(getShiftedEndDateTime(start, prevEnd, prevStartValue, false, localized)).to.equal(prevEnd);
+						});
+
+						it('should shift as expected when previously same and inclusive', () => {
+							const prevStartValue = '2020-10-25T06:00:00.000Z';
+							const prevEnd = '2020-10-25T06:00:00.000Z';
+							const start = '2020-10-25T08:00:00.000Z';
+							const newEndValue = `2020-10-25T08:00:00.000${localized ? '' : 'Z'}`;
+							expect(getShiftedEndDateTime(start, prevEnd, prevStartValue, true, localized)).to.equal(newEndValue);
+						});
+
+						it('should only shift at latest to 11:59 PM', () => {
+							let prevStartValue, prevEnd, start, newEndValue;
+							if (timezone === 'America/Toronto') {
+								prevStartValue = '2020-10-25T06:00:00.000Z';
+								prevEnd = '2020-10-25T20:00:00.000Z';
+								start = '2020-10-25T14:00:00.000Z';
+								newEndValue = localized ? '2020-10-25T23:59:00.000' : '2020-10-26T03:59:00.000Z';
+							} else {
+								prevStartValue = '2020-10-25T17:00:00.000Z';
+								prevEnd = '2020-10-25T23:00:00.000Z';
+								start = localized ? '2020-10-25T20:00:00.000Z' : '2020-10-26T10:15:00.000Z';
+								newEndValue = localized ? '2020-10-25T23:59:00.000' : '2020-10-26T15:14:00.000Z';
+							}
+
+							expect(getShiftedEndDateTime(start, prevEnd, prevStartValue, inclusive, localized)).to.equal(newEndValue);
+						});
+					});
+				});
+			});
+		});
 	});
 
 	describe('values', () => {
